Combine codefixes project into analyzer project
