--- RoslynAnalyzerTemplate.CodeFixes/RoslynAnalyzerTemplate.CodeFixes.csproj
@@ -1,22 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <IsPackable>false</IsPackable>
-    <IsRoslynComponent>true</IsRoslynComponent>
-    <RootNamespace>RoslynAnalyzerTemplate</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.3.1"/>
-    <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.3">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\RoslynAnalyzerTemplate\RoslynAnalyzerTemplate.csproj"/>
-  </ItemGroup>
-
-</Project>
