+++ RoslynAnalyzerTemplate.Test/RoslynAnalyzerTemplate.Test.csproj
@@ -9,24 +9,23 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0"/>
-    <PackageReference Include="Microsoft.CodeAnalysis" Version="4.3.1"/>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Analyzer.Testing.NUnit" Version="1.1.1"/>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.CodeFix.Testing.NUnit" Version="1.1.1"/>
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.CodeRefactoring.Testing.NUnit" Version="1.1.1"/>
-    <PackageReference Include="NUnit" Version="3.13.3"/>
-    <PackageReference Include="NUnit3TestAdapter" Version="4.3.0"/>
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.4.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis" Version="4.3.1" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.Analyzer.Testing.NUnit" Version="1.1.1" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.CodeFix.Testing.NUnit" Version="1.1.1" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp.CodeRefactoring.Testing.NUnit" Version="1.1.1" />
+    <PackageReference Include="NUnit" Version="3.13.3" />
+    <PackageReference Include="NUnit3TestAdapter" Version="4.3.0" />
     <PackageReference Include="NUnit.Analyzers" Version="3.5.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Dena.CodeAnalysis.Testing" Version="3.0.0"/>
-    <PackageReference Include="JunitXml.TestLogger" Version="3.0.114"/>
+    <PackageReference Include="Dena.CodeAnalysis.Testing" Version="3.0.0" />
+    <PackageReference Include="JunitXml.TestLogger" Version="3.0.114" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\RoslynAnalyzerTemplate.CodeFixes\RoslynAnalyzerTemplate.CodeFixes.csproj"/>
-    <ProjectReference Include="..\RoslynAnalyzerTemplate\RoslynAnalyzerTemplate.csproj"/>
+    <ProjectReference Include="..\RoslynAnalyzerTemplate\RoslynAnalyzerTemplate.csproj" />
   </ItemGroup>
 
 </Project>
